requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-dotenv
httpx
beautifulsoup4
openai==0.28
requests
